2020-04-17修改部分接口为ResultFul，修改导入excel逻辑

+++ src/main/java/com/primeton/liuzhichao/demo/service/IMsgService.java
@@ -36,7 +36,7 @@ public interface IMsgService {
 	 * @param uid    当前用户的id
 	 * @return      
 	 */
-	Boolean updataMsgState(Long flage, Long uid);
+	Boolean updataMsgState(Long flage, Integer uid);
 	
 	/**
 	 * 查询所有用户列表
